+++ src/xer_parser/reader.py
@@ -85,7 +85,7 @@ class Reader:
     Examples
     --------
 
-    >>> from xerparser_dev.reader import Reader
+    >>> from xer_parser.reader import Reader
     >>> xer = Reader("myproject.xer")
     >>> for project in xer.projects:
     ...     print(project)